Move all loaders to corresponding folder under /vertical
