+++ src/main/java/ai/knowly/langtorch/loader/vertical/DocumentLoader.java
@@ -1,4 +1,7 @@
-package ai.knowly.langtorch.loader;
+package ai.knowly.langtorch.loader.vertical;
+
+import ai.knowly.langtorch.loader.LoadOption;
+import ai.knowly.langtorch.loader.Loader;
 
 /** Extension of Connector for reading documents. */
 public abstract class DocumentLoader<R extends LoadOption> extends Loader<String, R> {}
