+++ src/main/java/ai/knowly/langtorch/loader/vertical/markdown/MarkdownLoadOption.java
@@ -1,12 +1,12 @@
-package ai.knowly.langtorch.loader.pdf;
+package ai.knowly.langtorch.loader.vertical.markdown;
 
 import ai.knowly.langtorch.loader.LoadOption;
 import lombok.Builder;
 import lombok.Data;
 
-/** Implementation of ReadOption for PDF files. */
+/** Implementation of ReadOption for Markdown files. */
 @Data
 @Builder(toBuilder = true, setterPrefix = "set")
-public class PDFLoadOption implements LoadOption {
+public class MarkdownLoadOption implements LoadOption {
   private String filePath;
 }
