+++ src/main/java/ai/knowly/langtorch/loader/vertical/sql/SQLLoader.java
@@ -1,7 +1,7 @@
-package ai.knowly.langtorch.loader.sql;
+package ai.knowly.langtorch.loader.vertical.sql;
 
-import ai.knowly.langtorch.loader.Loader;
 import ai.knowly.langtorch.loader.LoadOption;
+import ai.knowly.langtorch.loader.Loader;
 
 /** SQL loader. */
 public abstract class SQLLoader<R extends LoadOption, S extends StorageObject>
