+++ src/main/java/ai/knowly/langtorch/loader/vertical/sql/StorageObject.java
@@ -1,4 +1,4 @@
-package ai.knowly.langtorch.loader.sql;
+package ai.knowly.langtorch.loader.vertical.sql;
 
 /** Shared interface for objects reading with SQL loader. */
 public interface StorageObject {}
